+++ bin/watchCancelCOASeatReservationAuthorizationQueue/index.ts
@@ -14,7 +14,7 @@ mongoose.connect(<string>process.env.MONGOLAB_URI, mongooseConnectionOptions);
 let count = 0;
 
 const MAX_NUBMER_OF_PARALLEL_TASKS = 10;
-const INTERVAL_MILLISECONDS = 250;
+const INTERVAL_MILLISECONDS = 1000;
 const queueAdapter = sskts.adapter.queue(mongoose.connection);
 
 setInterval(